backend/requirements.txt
# Core Frameworks and API
django>=5.1.0,<5.2.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pyyaml


# Database Driver for PostgreSQL
psycopg2-binary

# Background Processing (Celery)
celery[redis]==5.6.0

# HTTP Client for External AI API Calls
requests

# Environment Variables Management
django-environ

# Optional: To store Celery task results in the Django database
django-celery-results

#openai:The AI engine
openai

#
python-dotenv